requirements.txt
python-telegram-bot>=21.0,<23.0
yandex-music>=3.0,<4.0
